shells/ksh*: Fix ksh 1.0.5 Regression (issue #660) version 2
From the upstream patch description:

Patch version two. In mac_copy(), don't internally backslash-escape
a backslash in a glob pattern bracket expression. This also fixes
the #549 regression reintroduced by the previous patch, at least for
standard glob patterns. For this, we need to use the bracketexpr flag
(introduced as a copyto() local variable in 6c73c8c) in mac_copy(),
so we move it to the Mac_t struct, making it globally accessible.
Initialisation is automatic.