Mk/Scripts: Fix ports depends-list.sh
depends-list.sh fails in two ways: one is a simple quoting bug which
causes empty FLAVOR variables to be propagated; the other is that
de-duplication is being done without regard for flavors.

The upshot of these is that where a port depends on a specific flavor
(first bug) or on multiple flavors (second bug, usually caused by build
dependencies) of another port, the constructed dependency list is
incomplete, and so -recursive targets do not do all the required work.
(This is especially annoying for fetch-recursive.)

Reviewed by: portmgr@FreeBSD.org
PR:		269982
Differential Revision: https://reviews.freebsd.org/D40491