graphics/podofo: update to 0.10.0 preserving version 0.9.x
Version 0.10.0 of podofo is API/ABI incompatible with previous
version. so with the update we need to preserve the previous version
in the tree as graphics/podofo09 for ports depending on that version.
The old port is modified so that it can be installed concurrently
with the new version.

Ports depending on the old version on the ports are modified
accordingly.

Approved by:		sunpoet (maintainer, via private mail), pkubaj, yuri
Differential Revision:	https://reviews.freebsd.org/D40328