sysutils/tracker3: add SOUP2 and SOUP3 multi options
Because Tracker can support libsoup2 and libsoup3 backend, and also at
the same time, it makes sense to add a multi option to give the user the
possibility of selecting them.

The default is still libsoup2, so the default dependencies of this port
will not change at this time.

PR:		272017
MFH:		2023Q2