devel/py-setuptools: fix CVE-2022-40897 backporting a patch
This commit integrates one-line upstream fix for the problem:
https://github.com/pypa/setuptools/commit/43a9c9bfa6aa626ec2a22540bea28d2ca77964be.diff

Our port has not been fixed for several months making users unhappy.
It's upto the maintainer to update the port, this commit does not update it.

Bump PORTREVISION and adjust VuXML entry.
Due to the nature of the problem and fix there is no need in updating consumers.