devel/orc: work around armv7 libc bug
The armv7 libc has a bug in printf() where printing a denormal double
may cause NUL bytes to be written out.  This port tries to write a
diagnostic comment containing such a denormal to a C source file.  The
bug causes the source file to be malformed, breaking the port and its
consumers.

Work around the bug by killing the offending part of the diagnostic
comment.

Approved by:	portmgr (build fix blanket)
Reviewed by:	diizzy
PR:		272229