converters/dos2unix: Add new port
This is a port for Erwin Waterlander's dos2unix, which is shipped in
most Linux distributions, and has many more features than the similar
converters/dosunix and converters/unix2dos ports. It is also much more
maintained, as those other ports have not had significant updates in
years.

The port has DOCS and NLS options, to toggle the building and/or
installing of documentation, multi-lingual messages and man pages.

Note that this port's "dos2unix" and "unix2dos" binaries conflict with
those of the converters/unix2dos port, but I see no good reason to have
both installed at the same time.

PR:		271998
Reviewed by:	eduardo
MFH:		2023Q2