ports-mgmt/pkg: update to 1.19.2
This release only deals with RSA signature for people using the PUBKEY
mechanism.
With openssl3 the signature had to change, in this release we make pkg
accept both the new signatures and the old ones.

if built with openssl 1 (aka freebsd 12 and 13) pkg will continue issueing
the usual signatures aka no changes.
if built with openssl 3 (aka freebsd 14) pkg will issue a new format of
signature.

The version of pkg older than 1.19.2 will not be able to use a signature
issued witg pkg running on freebsd 14

All system running 1.19.2 and above will be able to accept any of the
signatures.

If a user get stuck, for any reason, pkg bootstrap -f will fix the
situation as long as it installs 1.19.2 and above.

Note that is only concens PUBKEY which is not the default.

No changes at all for users of the official FreeBSD repositories.