lang/php80: Mark IGNORE_SSL for 14
This was marked IGNORE_SSL for OpenSSL 3 and later in an earlier commit
as there is no upstream support for OpenSSL 3 however that was reverted
as it got delayed in merging the OpenSSL in the base. As we have a
proper OSVERSION now this patch fixes the OSVERSION logic.

Sponsored by:	Bounce Experts
Sponsored by:	The FreeBSD Foundation