devel/lua-language-server: Deprecate and set to expire on 8/15/2023
This port is the most important lua LSP server, and is a huge
quality-of-life improvement for Neovim configuration.

However, the build process is bizarre and complex, and debugging
is very difficult. The tip of the iceberg is that the luamake
dependency comes only with a raw ninja file; that changes nothing,
but does represent the complexity of the the build process. Luamake
does nominally support FreeBSD, but definitely isn't tested on it.

Versions of LuaLS after 3.6.9 dump core immediately when used in
Neovim, and I don't have the time or expertise to figure out why.
Additionally, the last couple versions don't even compile.

If someone is inclined to debug LuaLS and get it working, I would
be absolutely thrilled to un-deprecate it. But if that doesn't
occur, it's not worth keeping in the tree.

The port is set to expire on 8/15/2023.