lang/perl5*: Link with -pthread instead of -lpthread
During configure of security/p5-Crypt-GCrypt a test program is
compiled and run to test libgcrypt.  This program segfaults when perl5
is built with the THREADS option.  Libgcrypt links with libgpg-error
which has an init function which calls into libintl (gettext) which
calls pthread functions before libthr is initialised.

The compiler command line is provided by perl and it includes both
-pthread and -lpthread.  Patch perl to remove -lpthread.  With -pthread
the compiler will pass -lpthread to the linker such that it is
initialised early.

Discovered by the exp-run for gettext 0.22.

While here also remove -lc unconditionally and remove the code that
replaced it with -lc_r for old version of FreeBSD because the version
test is incorrect (missing a dot).  The -pthread flag already tells
the compiler to use -lc_r on those old versions.

PR:		272210