sysutils/scanbd: correct file placement and robustness fix
These are additional revisions based on discussion on bug #272066.
There, the placement for the sample scripts wasn't correct. So, we now
move the sample scripts to $LOCALBASE/etc/scanbd/scripts as samples, but
they aren't marked as samples in the plist. There is also a pkg-message
that explains this.

This should be the last revision I can see in the immediate future until
a new release comes out.

PR:		272419
See also:	272066