games/qudos: try to unbreak the port's build against Clang 16
While here, remove compiler flags `-fomit-frame-pointer' and
`-funroll-loops': the effect of the former is minuscule these
days (but it still breaks debugging) and for the latter, one
should let the [modern] compiler decide which loops to unroll
and when; forcing it will most likely yield larger and slower
code due to instruction cache misses.

Last but not least, make some private functions in the OpenGL
renderer static and define LICENSE (GPLv2 or later).

Reported by:	pkg-fallout