emulators/wine-devel: Remove the STAGING option
The Wine Staging patchset is experimental and goes on top of bi-weekly
development snapshots of Wine on top of a non-standard (in the world of
Wine) operating system. That's a lot of moving parts and has proven too
much to reasonably handle and user uptake has been limited at best.

Plus upstream has completely reinvented the mechanism to apply this
patchset and adjusting to that (and the additional dependencies) would
further slow down finally upgrading this port. [1]

(files/extra-gitapply was about the previous mechanism around Wine
Staging.)

PR:		271201 [1]
Discussed with:	Damjan Jovanovic <damjan.jov@gmail.org>