security/sssd: Fix pkg-plist and chase krb5 update
This patch fixes two issues.

The first issue was presence of missing files of the SSSDConfig Python
module: PKG-INFO, SOURCES.txt, dependency_links., and top_level.txt. It
looks like those files are not installed anymore so let's just remove
them from pkg-plist.

The other issue was the build system not accepting the new version of
secruity/krb5. The fix is to patch sssd's build system to accept the new
version. This is how we fixed previous incarnations of this bug in the
past.

PR:		271939
Reported by:	ddrinnon@cdor.net
Fixes:	7228a09df016 security/krb5: krb5-121 is now default
MFH:		2023Q3
Sponsored by:	Klara Inc.