databases/akonadi: and net-im/kaccounts-integration: reverse dependencies
Without this patch, there is a circular dependency if Akonadi is built with
kaccounts-integration installed, as Akonadi has an optional dependency on
kaccounts-integration and libaccounts-qt5 that makes the port silently depend
on libkaccounts.so and libaccounts-qt5.so. kaccounts-integration, on the
other hand, had a dependency on Akonadi that was probably inherited from the
Qt4 version, as the Akonadi plugin was removed from the code base in 2015.

Reviewed by:		rakuco, tcberner
Tested by:		tcberner
Approved by:		tcberner (mentor, kde@)
Differential Revision:	https://reviews.freebsd.org/D41019