net-im/libquotient: Fix the build on FreeBSD releases with LLVM < 14
Older libc++ releases have a bug, https://reviews.llvm.org/D107199, that
prevents `connectSingleShot()` in this port from working in certain
cases. The current workaround was to just avoid compiling the test that
caused the build to fail, but it does not fix the problem for consumer
ports like net-im/neochat.

The easiest solution in this case is to make the port fall back to a
code path that does not use `std::bind_front()` on FreeBSD versions with
older LLVM releases.

Test Plan:
The port built fine on 12.4-i386 (LLVM 13.0.0), 13.2-amd64 (LLVM 14.0.3)
and 14-amd64 (LLVM 16). I didn't bother to test 13.1 because it's going
EOL at the end of the month anyway.

Approved by:		kde (arrowd)
MFH:			2023Q3
Differential Revision:	https://reviews.freebsd.org/D41044