net-im/libquotient: Fix logic in patch-lib_qt__connection__util.h
Follow-up to commit e7f65a58f219 ("net-im/libquotient: Fix the build on
FreeBSD releases with LLVM < 14"). The `#ifdef` logic added there was
slightly wrong, and we were basically disabling that code path even when
it was supposed to work.

Add a missing set of parentheses to fix the issue.

This has been tested on 12.4-i386, 13.2-amd64 and 14-amd64.

MFH:		2023Q3