graphics/osgearth: Unbreak build
The build failed with

  [...] error: no matching function for call to 'lerc_decode'
  /usr/local/include/Lerc_c_api.h:223:17: note: candidate function not
  viable: requires 10 arguments, but 9 were provided

and two similar errors. It obviously picked up lerc headers from
graphics/lerc although it uses a bundled lerc (which is directly built
in, not linked).

Add a patch to make sure the include path to the bundled lerc is
prepended.

With llvm15+, there was a second issue:

  /usr/include/c++/v1/__random/uniform_int_distribution.h:234:5: error:
  static assertion failed due to requirement
  '__libcpp_random_is_valid_urng<Catch::RandomNumberGenerator,
  void>::value'
  [...]
  [...]work/osgearth-osgearth-3.2/src/osgEarth/catch.hpp:6539:18: note:
  in instantiation of function template specialization
  'std::shuffle<std::__wrap_iter<Catch::TestCase *>,
  Catch::RandomNumberGenerator &>' requested here
              std::shuffle( vector.begin(), vector.end(), rng );

This is caused by the RNG used there having a signed return type
(ptrdiff_t) which is invalid, so add another patch to use size_t
instead.

Also update LIB_DEPENDS:
- remove poco, it was never used, the application using it isn't
  connected to the build
- remove leveldb, it's used by an optional component not enabled by
  default (could be supported by a port option)
- add webp, as advised by stage-qa

Approved by:	portmgr (blanket)