graphics/digikam: Update to 8.1.0
Remove DOCS option: The docs haven't been shipped with digikam starting
with 6.0.0 and are now in a separate repository:
https://invent.kde.org/documentation/digikam-doc
This could potentially be a separate port, but they've been gone for
4 years and the online docs are still available from within digikam
through the help menu or using a webbrowser.

Remove DOXYGEN option: Enabling the DOXYGEN option was just creating
the target to build the API docs, but we weren't actually building that
target. They probably are not useful to most users and would just
add to the build overhead.

Remove KVKONTAKTE option: Plugin was removed upstream.

Add FACEDETECT option: This installs data required for the facial
recognition feature to work. The files total about 420MB, so this is
set to off by default for now. Could be a seperate port to make it
easier to allow users to add later if they want.

Re-enable the MEDIAPLAYER option and add options pertaining to it.
QtAV was imported into the digikam codebase after being abandoned by
the original author and can now build with ffmpeg 6.x.

Add options for finer database control and disable internal MySQL
support by default as it only works with MariaDB.

8.1.0: https://www.digikam.org/news/2023-07-09-8.1.0_release_announcement/
8.0.0: https://www.digikam.org/news/2023-04-16-8.0.0_release_announcement/