KDE: mark the telepathy components deprecated
KDE will no longer ship the KDE telepathy components (ktp) as part
of the KDE Gear releases starting from the upcoming KDE Gear 23.08
this august [1], [2].

The ports will stay at their current version until removal at the
end of the quarter.

These components are not actively maintained, and the same goes
for the rest of the telepathy stack. [3] gives a quick overview,
explaining that it is mostly broken.

[1] https://invent.kde.org/sysadmin/release-tools/-/commit/9e8b75d787a33405a68379364d9637e7e7001328
[2] https://mail.kde.org/pipermail/release-team/2023-June/013068.html
[3] https://mail.kde.org/pipermail/release-team/2023-June/013080.html