devel/bloomberg-bde: new port for Bloomberg's foundational C++ libraries
BDE (Bloomberg Development Environment) is a collection of
libraries used inside Bloomberg. It's a bit like Boost, in
that it adds a whole spectrum of useful things to C++.
It's also like Boost in that things do flow from the library
into the language standard. BDE is Apache 2.0 licensed
and tracks modern C++ usage.

(This initial commit gets a library-that-builds into the
ports tree; it has no consumers yet so I won't swear that
it works completely)