devel/tig: Update to 2.5.8
Apart from the update, the patch also fixes a missing dependency.  If
pcre2 (or pcre) is detected during 'configure', it will enable the pcre
library during the build.

If so, 'make stage-qa' will complain:

> Error: /usr/local/bin/tig is linked to /usr/local/lib/libpcre2-posix.so.3 from devel/pcre2 but it is not declared as a dependency
> Warning: you need LIB_DEPENDS+=libpcre2-posix.so:devel/pcre2

The patch also adds a PCRE option that explicity enables / disables pcre
as desired.  The default state of the PCRE option is off which is just
like how packages are build in a package builder environment where the
pcre package is not installed.  But I have no problem with having it on
by default (committer's discretion).  The advantage is that it allows
users to search via regexp which is a useful tool.  The pcre libs have a
friendly BSD3CLAUSE license.

While here, tweak some sorting / formatting as hinted by
portfmt/portclippy.

PR:	270362