emulators/open-vm-tools: Add DEPLOYPKG option
Before this change, open-vm-tools port used to depend inconditionally of
libmspack, which is a requirement of DeployPkg plugin.  It turns out
that configure script only enable this plugin automatically on linux, so
it was not being built and dependency was not supposed to be there.

Instead of removing bogus dependency, I decided to add DEPLOYPKG option,
used to enable that plugin and left it off by default in order to
preserve previous default behavior.

Some changes on source code were needed to make DeployPkg to build on
FreeBSD and are done by new patches, which I intend to submit upstream.

Reported by:	vangyzen
Sponsored by:	Rubicon Communications, LLC ("Netgate")