biology/salmon: Transcript-level quantification of RNA-seq
Salmon is a wicked-fast program to produce a highly-accurate,
transcript-level quantification estimates from RNA-seq data. Salmon
achieves is accuracy and speed via a number of different innovations,
including the use of quasi-mapping (accurate but fast-to-compute
proxies for traditional read alignments), and massively-parallel
stochastic collapsed variational inference. The result is a versatile
tool that fits nicely into many differnt pipelines.