framework: add Mk/Uses/pycryptography
With the added new DEFAULT_VERSION "PYCRYPTOGRAHY_DEFAULT=rust|legacy"
it is important to transparently depend on the right one.

This is the job of this new uses.

So instead of manually adding dependency lines like

	RUN_DEPENDS=${PYTHON_PKGNAMEPREFIX}cryptography>0:security/py-cryptography@${PY_FLAVOR}

this should now be

	USES=pycryptography:run

Supported arguments: <none>,build,run,test
If no argument is given, it defaults to 'build,run'.

A future commit will update to security/py-cryptography will introduce a
rust dependency.

PR:		254853