devel/cxxtest: take maintainership, add docs/examples
This port is in need of a new maintainer, Hans Petter Selasky who was
the previous maintainer has passed away in June 2023.
https://lists.freebsd.org/archives/freebsd-announce/2023-July/000076.html

We thank Hans Petter for maintaining this port and his other contributions,
which have been countless.

I am taking maintainership of this port, which is
Approved by:	portmgr@ (discussed with bofh@ → blanket)

Apparently the home page is currently unavailable due to configuration
issues (no DNS A nor AAAA RR), so add a different WWW for now.
Issue reported upstream to https://github.com/CxxTest/cxxtest/issues/157

Add DOCS and EXAMPLES options, and install the relevant bits
(for docs: the guide.txt in asciidoc from the tarball,
and the .pdf downloaded separately from Github [1];
for examples: these are part of the distribution's tarball).

[1] NOTE: DOCS_DISTFILES needs an explicit DISTFILES. portlint does not know
this currently and will complain. We still need it, see the warning box at the end of
https://docs.freebsd.org/en/books/porters-handbook/makefiles/#options-variables

Bump PORTREVISION to make the DOCS/EXAMPLES addition effective.