graphics/osgearth: Unbreak build on arm64
The build failed on arm64 for including llvm intrinsics headers for x86
architectures. The extra patch for i386 already modified the build not
to use these intrinsics, so it solves the issue for arm64 as well. It is
still needed on i386 because intrinsics requiring SSE are used.

Therefore, rename the patch file to better describe the purpose and use
it for any architecture except amd64 and powerpc64, the latter assuming
the flags used for powerpc64 still work, which I cannot test myself.

The result on architectures that built fine before is unchanged,
therefore no revision bump is required.

Reported by:	pkg-fallout