games/quake2-3zb2: try to unbreak the port against Clang 16
The treatment is basically the same as in commit 2ca9bcd573e3
(-Wno-incompatible-function-pointer-types); using the MAKEFILE
instead of copying the Makefile to ${WRKSRC} is also alike.

While I'm here, prune the header files from the SRCS list and
banal comment in the `post-extract' target, use option helpers
during the installation, and stop messing with permissions as
it breaks ``make clean''.

Reported by:	pkg-fallout