net/rabbitmq: Fix `rabbitmq` rc.d script to use `%%_ERLANG_LIB%%`
[Why]
In commit c9033971719dc098b34276f11b8e5a5e22b966fd, the variable was
renamed from `ERLANG_LIB` to `_ERLANG_LIB`. However the rc.d script was
not updated to reflect that change.

Therefore the script didn't use the expected version of Erlang. This
meant that it could use `lang/erlang` if it was installed (which could
be incompatible) or not start at all.

[How]
The rc.d script is simply updated to use the new variable name. The port
revision is bumped to note the change.

PR:		271269
Reported by:	Miroslav Lachman <000.fbsd@quip.cz>