Mk/bsd.sites.mk: Add new MACRO TEX
There are two different sites/mirror required for the ports tree. Due to
the packaging of texlive we need the CTAN historic mirror which is
defined in TEX_CTAN and normal mirror which hosts most of the packages.
Unfortunately both were uses interchangeably without checking the
details and in previous there was single sites mirroring both the
historic dists and the noemal packages. As this is no longer the case
update all other ports pointing towards the wrong site.

Approved by:	portmgr (blanket)