x11-wm/fvwm3: Use upstream .desktop file
After upstream accepted a simple .desktop file with a few changes, use
this one until the next upstream version is released.

The main difference is that it uses Type=Application instead of
Type=XSession. According to freedesktop.org specs, Type=XSession does
not exist. Both work fine with sddm, but it's still better to use the
one that conforms to the spec.

PR:		272836