devel/qt6-base: Re-enable AVX2 after d2283436a461
The original code incorrectly assumes that all CPUs with FMA instruction
sets also have AVX2 instruction sets. This is true for Intel CPUs, but
not for AMD bdver[23] which have FMA, but not AVX2. All Intel and AMD
CPUs that have AVX2 also have BMI2, so use that as a check instead of
disabling newer instruction sets entirely.