security/R-cran-digest: fix build on armv7
Disable HAVE_MM_PREFETCH in the bundled crc32c as this option is
available on x86 only.  In any way, not using _mm_prefetch should not
have a significant performance impact as an arch-agnostic prefetch
method is used instead.

Upstream made the same change in 0.6.33, which is what we ship in the
main branch.

Approved by:	portmgr (build fix blanket)