www/angie-module-brotli: Angie Brotli dynamic module
Module adds Brotli compression support to Angie.
Brotli is a generic-purpose lossless compression algorithm that compresses
data using a combination of a modern variant of the LZ77 algorithm, Huffman
coding and 2nd order context modeling, with a compression ratio comparable
to the best currently available general-purpose compression methods.

It is similar in speed with deflate but offers more dense compression.

PR:		272999