x11/antimicrox: New port to map gamepad keys to keyboard, mouse, scripts and macros
AntiMicroX is a graphical program used to map gamepad keys to keyboard, mouse, scripts
and macros. You can use this program to control any desktop application with a gamepad
on Linux, Windows and FreeBSD.
It can be also used for generating SDL2 configuration (useful for mapping atypical
gamepads to generic ones like xbox360).

This port is a suggested replacement for x11/antimicro which is unmaintained since 2017.