sysutils/cpu-microcode-*: Reorganize CPU microcode ports
Prior to this update, sysutils/devcpu-data, which only contained an RC
script, had run dependencies on the AMD and Intel microcode ports.  This
made it cumbersome to have just the AMD or just the Intel microcode
ports installed.  With this change, the microcode ports now depend on
the RC script.

Other changes:

- Use more intuitive port names: cpu-microcode, cpu-microcode-amd,
  cpu-microcode-intel, and cpu-microcode-rc.
- Add the metaport, cpu-microcode, which pulls in all related ports.
- Pet portclippy/portfmt

Reviewed by:	lwhsu, markj
Approved by:	sbruno (maintainer)
Sponsored by:	The FreeBSD Foundation
Differential Revision:	https://reviews.freebsd.org/D41406