textproc/py-interegular: Add py-interegular 0.3.2
Interegular is a library to check a subset of python regexes for intersections.
Based on grennery by @qntm. Adapted for lark-parser.

The primary difference with grennery library is that interegular is focused on
speed and compatibility with python re syntax, whereas grennery has a way to
reconstruct a regex from a FSM, which interegular lacks.