emulators/wine-devel: Add run-time dependency on alsa-plugins
When using the ALSA option (which is not the default) audio did not
work - no audio device appeared for playing and recording and the
following message appeared:

  ALSA lib dlmisc.c:285:(snd_dlobj_cache_get0) Cannot open shared
  library /usr/local/lib/alsa-lib/libasound_module_pcm_oss.so

Providing alsa-plugins at run time addresses this.

(This is a forward port of a26844393b and 3a2594ed6f from the primary
wine port.)

Bump PORTREVSION accordingly.

PR:		273137