devel/py-pykdtree: Add py-pykdtree 1.3.7.post0
pykdtree is a kd-tree implementation for fast nearest neighbour search in
Python. The aim is to be the fastest implementation around for common use cases
(low dimensions and low number of neighbours) for both tree construction and
queries.

The implementation is based on scipy.spatial.cKDTree and libANN by combining the
best features from both and focus on implementation efficiency.

The interface is similar to that of scipy.spatial.cKDTree except only Euclidean
distance measure is supported.

Queries are optionally multithreaded using OpenMP.