devel/rubygem-syntax_tree: Add rubygem-syntax_tree 6.1.1
Syntax Tree is a suite of tools built on top of the internal CRuby parser. It
provides the ability to generate a syntax tree from source, as well as the tools
necessary to inspect and manipulate that syntax tree. It can be used to build
formatters, linters, language servers, and more.

It is built with only standard library dependencies. It additionally ships with
a plugin system so that you can build your own syntax trees from other languages
and incorporate these tools.