devel/boost-all: Adjust range to use llvm from ports
When boost was updated to 1.81.0, the problem was reported as PR
269067 that build of devel/boost-libs fails because of assertion
failure of Clang in base system. According to the comment submitted to
upstream issue tracker (*), the source of the problem is very long
standing bug of LLVM/Clang. So the bug itself is included in all of
supported branch of src repository. But the problem only happens with
main because assertion is only enabled with it. To work it around
compiled.mk was modified so llvm from ports is used instead of that of
base system when __FreeBSD_version is 1400000 or later. This is what
was done with commit 3efdfab9b59f.

And recently stable/14 branch is created in src repository. After it
is created __FreeBSD_version of the branch is bumped from 1400096 to
1400097. And the value of main is bumped from 1400096 to 1500000. So
adjust range to use llvm from ports as following.

Case 1: 1400000 <= OSVERSION < 1400097 (main before stable/14 is branched)
Case 2: OSVERSION >= 1500000 (main after stable/14 is branched)

Reference:	https://github.com/llvm/llvm-project/issues/39319#issuecomment-1444452238 (*)
PR:		273335
Approved by:	sunpoet (office@, maintainer)
MFH:		2023Q3

(cherry picked from commit 9b2e4b0219bd4814b47c22e5901f3b1cc091fded)