security/ncrack: fix bad assumption in configure check
A version check in opensshlib/configure assumed that the zlib version
had 3 elements (e.g. 1.2.3), but the latest version is 1.3 (not 1.3.0).
Adjust the configure check to accept a 2-element version number.

No PORTREVISION bump, since this is just a fix for a build check that
caused an abort of the port build on recent -CURRENT and 14-ALPHA.