print/ghostscript10: add default port option for X11
This change replaces flavors with a new default port option for X11.
Compared to GS 9 port the X11 device driver is now integrated into GS instead
of being dynamically loaded which is not in conformance with new GS 10 approach.
The flavors approach creates conflicting dependencies with other ports which
cannot be solved this time with the ports framework.

See PRs 270989 and 272541 for details.

PR:		272541
Tested by:	kib