www/galene: new port
Galène (or Galene) is a videoconference server (an “SFU”) that is easy to
deploy and that requires moderate server resources. It was originally designed
for lectures, conferences and student tutorials, but is also useful for
traditional meetings.

Galène has been used in production at two major universities (Université de
Paris and Sorbonne Université) for lectures, practicals, seminars, and for
staff meetings. It has been used to host two conferences (SOCS'2020 and
JFLA'2021).